# docs/openapi/admin-api.yaml
openapi: 3.0.3
info:
  title: MujahidMusic Admin API
  version: 1.0.0
  description: BFF for admin-web (port 8080). Auth via Bearer ADMIN_API_KEY.
servers:
  - url: http://localhost:8080
security:
  - AdminBearer: []
paths:
...
  /api/admin/playlists:
    get:
      summary: List playlists with pagination
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 50
        - name: ownerId
          in: query
          schema:
            type: string
        - name: name
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated playlists
...
  /api/admin/playlists/{ownerId}/{name}:
    get:
      summary: Playlist tracks
      parameters:
        - name: ownerId
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Playlist detail
    delete:
      summary: Delete playlist
      parameters:
        - name: ownerId
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    AdminBearer:
      type: http
      scheme: bearer